requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
boto3==1.26.90
botocore==1.29.90
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-dump-load-utf8==0.0.4
django-import-export==3.1.0
django-rest-multiple-models==2.1.3
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.5.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tablib==3.3.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
